+++ doc/release_notes.rst
@@ -8,6 +8,10 @@ This list summarizes the changes between Humble (previous) and Jazzy (current) r
 
   This list was created in June 2025 (tag 3.6.0), earlier changes may not be included.
 
+AsyncFunctionHandler
+*******************************
+* Add DETACHED scheduling policy for Async function handler (`#383 <https://github.com/ros-controls/realtime_tools/pull/383>`__).
+
 RealtimeBox
 *******************************
 * ``RealtimeBox`` got renamed to ``RealtimeThreadSafeBox`` and uses real-time mutexes now (`#318 <https://github.com/ros-controls/realtime_tools/pull/318>`__, `#342 <https://github.com/ros-controls/realtime_tools/pull/342>`__).